Created FavoriteCollectionViewModel
