+++ MyPexels/FavoriteCollection/FavoriteCollectionViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  FavoriteCollectionViewModel.swift
+//  MyPexels
+//
+//  Created by Artem Pavlov on 23.10.2022.
+//
+
+import Foundation
