+++ MyPexels/TabBarView/TabBarStartViewController.swift
@@ -14,20 +14,17 @@ protocol TabBarStartViewControllerDelegate {
 class TabBarStartViewController: UITabBarController {
     
     //MARK: - Public Properties
-    var viewModel: TabBarStartViewModelProtocol!
+    private var viewModel: TabBarStartViewModelProtocol!
     
     //MARK: - Private Properties
     private let photosVC = PhotoCollectionViewController(collectionViewLayout: UICollectionViewFlowLayout())
     private let favoriteVC = FavoriteCollectionViewController(collectionViewLayout: UICollectionViewFlowLayout())
     private let settingsVC = SettingsViewController()
-    private var favoritePhotos: [PexelsPhoto] = []
     
     //MARK: - Life Cycles Methods
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        loadFavouritePhotos()
-        viewModel = TabBarStartViewModel(favoritePhotos: favoritePhotos)
+        viewModel = TabBarStartViewModel()
         setupTabBar()
         updateFavotiteData(for: photosVC)
         updateFavotiteData(for: favoriteVC)
@@ -52,16 +49,16 @@ class TabBarStartViewController: UITabBarController {
         viewControllers = [photosVC, favoriteVC, settingsVC]
     }
     
-    private func updateFavotiteData(for favoriteViewController: FavoriteCollectionViewController) {
-        favoriteViewController.favoritePhotos = favoritePhotos
-        favoriteViewController.delegateTabBarVC = self
-    }
-    
     private func updateFavotiteData(for photoViewController: PhotoCollectionViewController) {
-        
         photoViewController.viewModel = viewModel.photoCollectionViewModel()
         photoViewController.delegateFavoriteVC = favoriteVC
         photoViewController.delegateTabBarVC = self
+    }
+    
+    private func updateFavotiteData(for favoriteViewController: FavoriteCollectionViewController) {
+        //favoriteViewController.favoritePhotos = favoritePhotos
+        favoriteViewController.favoritePhotos = viewModel.getFavoritePhotos()
+        favoriteViewController.delegateTabBarVC = self
     }
     
     private func updateFavotiteData(for userViewController: SettingsViewController) {
@@ -74,23 +71,12 @@ class TabBarStartViewController: UITabBarController {
         title = tabBar.selectedItem?.title
         navigationController?.navigationBar.prefersLargeTitles = true
     }
-    
-    private func loadFavouritePhotos() {
-        StorageManager.shared.fetchFavoritePhotos { result in
-            switch result {
-            case .success(let photos):
-                self.favoritePhotos = photos
-            case .failure(let error):
-                print(error.localizedDescription)
-            }
-        }
-    }
 }
 
 //MARK: - TabBarStartViewControllerDelegate
 extension TabBarStartViewController: TabBarStartViewControllerDelegate {
     func reloadFavoriteData() {
-        loadFavouritePhotos()
+      //  loadFavouritePhotos()
         updateFavotiteData(for: favoriteVC)
         updateFavotiteData(for: photosVC)
     }
